# packages/backend-sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/invitations": {
      "get": {
        "operationId": "InvitationCoreController_findProjectInvitations",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationSerialized"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/invitations": {
      "post": {
        "operationId": "InvitationCoreController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/invitations/{id}": {
      "get": {
        "operationId": "InvitationCoreController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      },
      "delete": {
        "operationId": "InvitationCoreController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/invitations/{id}/accept": {
      "post": {
        "operationId": "InvitationCoreController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationSerialized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
...
    "/auth/cli-flow/sessions": {
      "post": {
        "operationId": "CliFlowController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCliSessionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartCliSessionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth (cli-flow)"
        ]
      }
    },
    "/auth/cli-flow/sessions/{sessionId}": {
      "get": {
        "operationId": "CliFlowController_info",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliSessionInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth (cli-flow)"
        ]
      }
    },
    "/auth/cli-flow/sessions/{sessionId}/approve": {
      "post": {
        "operationId": "CliFlowController_approve",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCliSessionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth (cli-flow)"
        ]
      }
    },
    "/auth/cli-flow/sessions/{sessionId}/poll": {
      "get": {
        "operationId": "CliFlowController_poll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollCliSessionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth (cli-flow)"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "RefreshTokenCoreController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth (refresh)"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "RefreshTokenCoreController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth (refresh)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cryptly",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "project-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "project-api-key"
      }
    },
    "schemas": {
...
      "UserPartialSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "publicKey": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "avatarUrl",
          "displayName"
        ]
      },
...
      "InvitationSerialized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/UserPartialSerialized"
          },
          "role": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "admin"
            ]
          },
          "temporaryPublicKey": {
            "type": "string"
          },
          "temporaryPrivateKey": {
            "type": "string"
          },
          "temporarySecretsKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "author",
          "role",
          "temporaryPublicKey",
          "temporaryPrivateKey",
          "temporarySecretsKey",
          "createdAt"
        ]
      },
      "CreateInvitationBody": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "temporaryPublicKey": {
            "type": "string"
          },
          "temporaryPrivateKey": {
            "type": "string"
          },
          "temporarySecretsKey": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "admin"
            ]
          }
        },
        "required": [
          "projectId",
          "temporaryPublicKey",
          "temporaryPrivateKey",
          "temporarySecretsKey",
          "role"
        ]
      },
      "AcceptInvitationBody": {
        "type": "object",
        "properties": {
          "newSecretsKey": {
            "type": "string"
          }
        },
        "required": [
          "newSecretsKey"
        ]
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "isNewUser": {
            "type": "boolean"
          }
        },
        "required": [
          "token",
          "refreshToken"
        ]
      },
...
      "StartCliSessionBody": {
        "type": "object",
        "properties": {
          "tempPublicKey": {
            "type": "string",
            "description": "Base64 SPKI of the CLI temp RSA-OAEP public key"
          },
          "deviceName": {
            "type": "string",
            "description": "Human-readable device label, e.g. \"macbook-air · darwin\""
          }
        },
        "required": [
          "tempPublicKey",
          "deviceName"
        ]
      },
      "StartCliSessionResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "approveUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "number",
            "description": "Unix milliseconds at which this session expires."
          }
        },
        "required": [
          "sessionId",
          "approveUrl",
          "expiresAt"
        ]
      },
      "CliSessionInfoResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "consumed"
            ]
          },
          "expiresAt": {
            "type": "number",
            "description": "Unix milliseconds at which this session expires."
          },
          "tempPublicKey": {
            "type": "string",
            "description": "Base64 SPKI of the CLI temp public key — needed by the browser to wrap the private key."
          }
        },
        "required": [
          "sessionId",
          "deviceName",
          "status",
          "expiresAt",
          "tempPublicKey"
        ]
      },
      "ApproveCliSessionBody": {
        "type": "object",
        "properties": {
          "wrappedKey": {
            "type": "string",
            "description": "RSA-OAEP-encrypted ephemeral AES key (wraps the user private key)"
          },
          "encryptedPrivateKey": {
            "type": "string",
            "description": "AES-GCM-encrypted user private key (PKCS8)"
          }
        },
        "required": [
          "wrappedKey",
          "encryptedPrivateKey"
        ]
      },
      "PollCliSessionResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "consumed"
            ]
          },
          "jwt": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "wrappedKey": {
            "type": "string",
            "description": "RSA-OAEP-encrypted ephemeral AES key"
          },
          "encryptedPrivateKey": {
            "type": "string",
            "description": "AES-GCM-encrypted user private key (PKCS8)"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "RefreshBody": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
    }
  }
}